#42: fixed max_workers and no-parallel
Also improved help information
