+++ crates/assemble-freight/Cargo.toml
@@ -18,7 +18,7 @@ categories = ["development-tools"]
 [dependencies]
 assemble-core = { path = "../assemble-core", version = "0.2.0" }
 backtrace = "0.3.65"
-clap = { version = "4.0.4", features = ["derive", "color"] }
+clap = { version = "4.0.4", features = ["derive", "color", "cargo", "help"] }
 num_cpus = "1.13.1"
 petgraph = "0.6.2"
 thiserror = "1.0.31"
