+++ crates/assemble-freight/src/core/cli.rs
@@ -17,13 +17,26 @@ use std::num::NonZeroUsize;
 use std::path::PathBuf;
 use std::str::FromStr;
 
-/// The args to run Freight
+/// Command line options for running assemble based projects.
+///
+/// Tasks can either be the full path for the task, or a relative one from the directory
+/// in use within the project.
+///
+/// Task options are configured on per task basis and are fully configured at
+/// compile time. Options for tasks must immediately follow the task request.
+///
+/// When many tasks are matched for the same task request, they all
+/// receive the same task options.
 #[derive(Debug, Parser, Clone)]
 #[clap(name = "assemble")]
-#[clap(about, version)]
+#[clap(version, author)]
+#[clap(before_help = format!("{} v{}", clap::crate_name!(), clap::crate_version!()))]
+#[clap(after_help = "For project specific information, use the :help task.")]
 #[clap(allow_hyphen_values = true)]
+#[clap(term_width = 64)]
 pub struct FreightArgs {
     /// Tasks to be run
+    #[clap(value_name = "TASK [TASK OPTIONS]...")]
     bare_task_requests: Vec<String>,
     /// Project lazy_evaluation. Set using -P or --prop
     #[clap(flatten)]
@@ -31,27 +44,35 @@ pub struct FreightArgs {
     /// Log level to run freight in.
     #[clap(flatten)]
     pub logging: LoggingArgs,
+
     /// The number of workers to use.
     ///
     /// Defaults to the number of cpus on the host.
     #[clap(long, short = 'J')]
     #[clap(default_value_t = NonZeroUsize::new(num_cpus::get()).expect("Number of cpus should never be 0"))]
-    #[clap(default_value_if("no-parallel", ArgPredicate::IsPresent, "1"))]
+    #[clap(default_value_if("no_parallel", ArgPredicate::IsPresent, "1"))]
+    #[clap(help_heading = None)]
     pub workers: NonZeroUsize,
     /// Don't run with parallel tasks
     #[clap(long)]
     #[clap(conflicts_with = "workers")]
+    #[clap(help_heading = None)]
     pub no_parallel: bool,
 
     /// Use an alternative settings file
     #[clap(short = 'F')]
+    #[clap(help_heading = None)]
     pub settings_file: Option<PathBuf>,
+
+
     /// Display backtraces for errors if possible.
     #[clap(short = 'B', long)]
+    #[clap(help_heading = None)]
     pub backtrace: bool,
 
     /// Forces all tasks to be rerun
     #[clap(long)]
+    #[clap(help_heading = None)]
     pub rerun_tasks: bool,
 }
 
@@ -77,6 +98,11 @@ impl FreightArgs {
         clone.bare_task_requests = iter.into_iter().map(|s| s.to_string()).collect();
         clone
     }
+
+    /// Gets a property.
+    pub fn property(&self, key: impl AsRef<str>) -> Option<&str> {
+        self.properties.property(key)
+    }
 }
 
 impl<S: AsRef<str>> FromIterator<S> for FreightArgs {
@@ -97,4 +123,47 @@ pub fn main_progress_bar_style(failing: bool) -> ProgressStyle {
     ProgressStyle::with_template(template)
         .unwrap()
         .progress_chars("=> ")
+}
+
+#[cfg(test)]
+mod test {
+    use super::*;
+    use crate::core::cli::FreightArgs;
+
+    #[test]
+    fn no_parallel() {
+        let args: FreightArgs = FreightArgs::command_line("--no-parallel");
+        println!("{:#?}", args);
+        assert!(args.no_parallel);
+        assert_eq!(args.workers.get(), 1);
+    }
+
+    #[test]
+    fn arbitrary_workers() {
+        let args: FreightArgs = FreightArgs::command_line("--workers 13");
+        println!("{:#?}", args);
+        assert_eq!(args.workers.get(), 13);
+        assert!(FreightArgs::try_parse_from(&["", "-J", "0"]).is_err());
+    }
+
+    #[test]
+    fn default_workers_is_num_cpus() {
+        let args: FreightArgs = FreightArgs::parse_from(&[""]);
+        assert_eq!(args.workers.get(), num_cpus::get());
+    }
+
+    #[test]
+    fn workers_and_no_parallel_conflicts() {
+        assert!(FreightArgs::try_parse_from(&["", "--workers", "12", "--no-parallel"]).is_err());
+    }
+
+    #[test]
+    fn can_set_project_properties() {
+        let args = FreightArgs::command_line("-P hello=world -P key1 -P key2");
+        assert_eq!(args.property("hello"), Some("world"));
+        assert_eq!(args.property("key1"), Some(""));
+        assert_eq!(args.property("key2"), Some(""));
+    }
+
+
 }
