+++ crates/assemble-freight/src/lib.rs
@@ -34,35 +34,3 @@ pub mod utils;
 
 pub use crate::core::cli::FreightArgs;
 
-#[cfg(test)]
-mod test {
-    use super::*;
-    use crate::core::cli::FreightArgs;
-
-    #[test]
-    fn no_parallel() {
-        let args: FreightArgs = FreightArgs::parse_from(&["", "--no-parallel"]);
-        println!("{:#?}", args);
-        assert!(args.no_parallel);
-        assert_eq!(args.workers.get(), 1);
-    }
-
-    #[test]
-    fn arbitrary_workers() {
-        let args: FreightArgs = FreightArgs::parse_from(&["", "--workers", "13"]);
-        println!("{:#?}", args);
-        assert_eq!(args.workers.get(), 13);
-        assert!(FreightArgs::try_parse_from(&["", "-J", "0"]).is_err());
-    }
-
-    #[test]
-    fn default_workers_is_num_cpus() {
-        let args: FreightArgs = FreightArgs::parse_from(&[""]);
-        assert_eq!(args.workers.get(), num_cpus::get());
-    }
-
-    #[test]
-    fn workers_and_no_parallel_conflicts() {
-        assert!(FreightArgs::try_parse_from(&["", "--workers", "12", "--no-parallel"]).is_err());
-    }
-}
